load VAE from file, fix likelihood computations

+++ run_vae.py
@@ -1,5 +1,6 @@
 from vae import VAE
 import pickle
+import os
 from utility import one_hot_encoding
 from vae import train, evaluate
 from pyro.infer import SVI, JitTrace_ELBO
@@ -59,8 +60,8 @@ if __name__ == "__main__":
     train_loader = torch.utils.data.DataLoader(seq_train, batch_size=batch_size)
     test_loader = torch.utils.data.DataLoader(seq_test, batch_size=batch_size)
 
-    wt = torch.tensor(one_hot_sequence[0])
-    seq_1 = torch.tensor(one_hot_sequence[1])
+    wt = torch.tensor(one_hot_sequence[0], dtype=torch.float)
+    seq_1 = torch.tensor(one_hot_sequence[1], dtype=torch.float)
 
     # parameters
     param_dict = {"LEARNING_RATE": args.learn_rate,
@@ -77,30 +78,34 @@ if __name__ == "__main__":
     experiment_name = args.experiment if args.experiment else f"VAE_Adam_z{args.latent_dim}"
     mlflow.set_experiment(experiment_name)
     print(f"Tracking URI: {mlflow.get_tracking_uri()}")
-    with mlflow.start_run():
-        vae = VAE(z_dim=param_dict["LATENT_DIM"], hidden_dim=param_dict["HIDDEN_DIM"], 
+    model_FILENAME = f"./models/VAE_{args.latent_dim}_{args.hidden_dim}_{args.epochs}.pt"
+    optimizer_FILENAME = f"./model/Adam_{args.latent_dim}_{args.hidden_dim}_{args.epochs}.pt"
+    vae = VAE(z_dim=param_dict["LATENT_DIM"], hidden_dim=param_dict["HIDDEN_DIM"], 
                     input_dims=param_dict["INPUT_DIM"], use_cuda=args.cuda, wt=wt)
-        optimizer = Adam({"lr": param_dict["LEARNING_RATE"]})
-        #optimizer = ClippedAdam({"lr": LEARNING_RATE})
-        svi = SVI(vae.model, vae.guide, optimizer, loss=JitTrace_ELBO())
+    optimizer = Adam({"lr": param_dict["LEARNING_RATE"]})
+    #optimizer = ClippedAdam({"lr": LEARNING_RATE})
+    svi = SVI(vae.model, vae.guide, optimizer, loss=JitTrace_ELBO())
+    if os.path.exists(model_FILENAME):
+        vae.load_state_dict(torch.load(model_FILENAME))
+    else:
+        with mlflow.start_run():
+            mlflow.log_params(param_dict)
+            for epoch in tqdm(range(args.epochs)):
+                total_epoch_loss_train = train(svi, train_loader, args.cuda)
+                mlflow.log_metric(key="neg loss train", value=-total_epoch_loss_train, 
+                                    step=epoch)
+                print(f"[epoch {epoch}] avrg. train loss: {total_epoch_loss_train}")
 
-        mlflow.log_params(param_dict)
-
-        for epoch in tqdm(range(args.epochs)):
-            total_epoch_loss_train = train(svi, train_loader, args.cuda)
-            mlflow.log_metric(key="neg loss train", value=-total_epoch_loss_train, step=epoch)
-            print(f"[epoch {epoch}] avrg. train loss: {total_epoch_loss_train}")
-
-            if epoch % args.validate == 0:
-                total_epoch_loss_test = evaluate(svi, test_loader, args.cuda)
-                mlflow.log_metric(key="neg loss validate", value=-total_epoch_loss_test, step=epoch)
-                print(f"[epoch {epoch}] avrg. test loss: {total_epoch_loss_test}")
-        
-        model_FILENAME = f"./models/VAE_{args.latent_dim}_{args.hidden_dim}_{args.epochs}"
-        optimizer_FILENAME = f"./model/Adam_{args.latent_dim}_{args.hidden_dim}_{args.epochs}"
-        torch.save(vae.state_dict(), model_FILENAME)
-        mlflow.log_artifact(model_FILENAME)
-        print(vae.likelihood(wt))
-        print(vae.log_odd_ratio(seq_1))
-        #torch.save(optimizer.state_dict(), optimizer_FILENAME)
-        #mlflow.log_artifact(optimizer_FILENAME)
+                if epoch % args.validate == 0:
+                    total_epoch_loss_test = evaluate(svi, test_loader, args.cuda)
+                    mlflow.log_metric(key="neg loss validate", value=-total_epoch_loss_test, 
+                                        step=epoch)
+                    print(f"[epoch {epoch}] avrg. test loss: {total_epoch_loss_test}")
+            
+            
+            torch.save(vae.state_dict(), model_FILENAME)
+            mlflow.log_artifact(model_FILENAME)
+            print(vae.likelihood(wt))
+            print(vae.log_odd_ratio(seq_1))
+            #torch.save(optimizer.state_dict(), optimizer_FILENAME)
+            #mlflow.log_artifact(optimizer_FILENAME)
